refactor organization skill actions
