+++ src/routes/Skills/OrganizationManagement.tsx
@@ -12,7 +12,6 @@ import type {
   BusyAction,
   LoadState,
   MemberView,
-  OrganizationSkillLinkInput,
   ProviderAccessForm,
   ProviderGrantView,
 } from "./organization-management-model.ts"
@@ -75,13 +74,7 @@ import {
   OrganizationSkillManageLoadingSkeleton,
   RuntimeSkillRemoveConfirmDialog,
 } from "./OrganizationSkillManageDialog.tsx"
-import { useSkillService } from "@/components/AppContext"
-import {
-  useAuthStateResource,
-  useHomeSummaryResource,
-  useSkillInventoryResource,
-  useSkillVersionReportResource,
-} from "@/components/AppDataHooks"
+import { useAuthStateResource, useSkillInventoryResource } from "@/components/AppDataHooks"
 import { Badge } from "@/components/ui/badge"
 import { Button } from "@/components/ui/button"
 import {
@@ -111,13 +104,10 @@ import {
 import { cn } from "@/lib/utils"
 import { useProviderSkillPackageLookup } from "@/routes/Skills/provider-skill-package-lookup"
 import { buildProviderSkillRecommendations } from "@/routes/Skills/provider-skill-recommendations"
-import {
-  canInstallPublicSkill,
-  getOrganizationSkillRuntimeStatus,
-  getPublicPackagePrimarySkill,
-} from "@/routes/Skills/skill-route-model"
+import { canInstallPublicSkill, getOrganizationSkillRuntimeStatus } from "@/routes/Skills/skill-route-model"
 import { useOrganizationAvatarPreviews } from "@/routes/Skills/use-organization-avatar-previews"
 import { useOrganizationMemberSearch } from "@/routes/Skills/use-organization-member-search"
+import { useOrganizationSkillActions } from "@/routes/Skills/use-organization-skill-actions"
 
 type AsyncResult<T> = { ok: true; value: T } | { error: unknown; ok: false }
 
@@ -126,20 +116,6 @@ function settle<T>(promise: Promise<T>): Promise<AsyncResult<T>> {
     (value) => ({ ok: true, value }),
     (error: unknown) => ({ error, ok: false }),
   )
-}
-
-function publicPackageLinkInput(pkg: PublicSkillPackage, skillName?: string): OrganizationSkillLinkInput | null {
-  const skill = skillName
-    ? (pkg.skills.find((item) => item.name === skillName) ?? getPublicPackagePrimarySkill(pkg))
-    : getPublicPackagePrimarySkill(pkg)
-  if (!skill) {
-    return null
-  }
-  return {
-    packageName: pkg.name,
-    skillName: skill.name,
-    version: pkg.version,
-  }
 }
 
 function mergeOrganizationUpdate(current: Organization, updated: Organization): Organization {
@@ -181,11 +157,8 @@ export function OrganizationManagementRoute({
   workspace?: UseOrganizationWorkspace
 }) {
   const { t } = useAppI18n()
-  const skillService = useSkillService()
   const authResource = useAuthStateResource()
   const skillInventory = useSkillInventoryResource()
-  const skillVersionReport = useSkillVersionReportResource()
-  const homeSummaryResource = useHomeSummaryResource()
   const activeAccount = authResource.data?.status === "authenticated" ? authResource.data.account : undefined
   const activeAccountId = activeAccount?.id
   const activeWorkspace = workspace?.activeWorkspace
@@ -227,7 +200,6 @@ export function OrganizationManagementRoute({
   const [addMemberOpen, setAddMemberOpen] = React.useState(false)
   const [membersPanelOpen, setMembersPanelOpen] = React.useState(false)
   const [providerAccessForm, setProviderAccessForm] = React.useState<ProviderAccessForm>(initialProviderAccessForm)
-  const [runtimeSkillRemoveTarget, setRuntimeSkillRemoveTarget] = React.useState<RuntimeSkillRemoveTarget | null>(null)
   const overviewRequestId = React.useRef(0)
   const detailsRequestId = React.useRef(0)
   const editAvatarUploadVersion = React.useRef(0)
@@ -256,6 +228,20 @@ export function OrganizationManagementRoute({
   }, [editOrganizationId, organizations])
   const selectedOrganizationSkills =
     selectedOrganization && organizationSkills?.organizationId === selectedOrganization.id ? organizationSkills : null
+  const {
+    addOrganizationSkillBatch,
+    addOrganizationSkillFromPackage,
+    addOrganizationSkillFromRecommendation,
+    installRuntimeSkill,
+    installRuntimeSkills,
+    removeRuntimeSkill,
+    runtimeSkillRemoveTarget,
+    setRuntimeSkillRemoveTarget,
+  } = useOrganizationSkillActions({
+    busyAction,
+    organizationSkills: selectedOrganizationSkills,
+    setBusyAction,
+  })
   const skillGroupById = React.useMemo(
     () => new Map((skillInventory.data?.groups ?? []).map((group) => [group.id, group])),
     [skillInventory.data?.groups],
@@ -989,223 +975,6 @@ export function OrganizationManagementRoute({
       }
     },
     [canManage, providerAccessError, selectedOrganization, t],
-  )
-
-  const installRuntimeSkill = React.useCallback(
-    async (skill: { packageName: string; skillName: string }) => {
-      setBusyAction(`installSkill:${skill.packageName}:${skill.skillName}`)
-      try {
-        const nextInventory = await skillService.invoke("installRegistrySkill", {
-          packageName: skill.packageName,
-          skillId: skill.skillName,
-        })
-        skillInventory.setData(nextInventory)
-        homeSummaryResource.invalidate()
-        toast.success(t("skills.registryInstallDone", { name: skill.skillName }))
-      } catch (error) {
-        toast.error(t("skills.registryInstallFailed", { error: errorMessage(error) }))
-      } finally {
-        setBusyAction(null)
-      }
-    },
-    [homeSummaryResource, skillInventory, skillService, t],
-  )
-
-  const removeRuntimeSkill = React.useCallback(async () => {
-    const target = runtimeSkillRemoveTarget
-    if (!target || busyAction) {
-      return
-    }
-
-    setBusyAction(runtimeSkillRemoveBusyKey(target))
-    try {
-      const nextInventory = await skillService.invoke("deleteSkill", {
-        confirmed: true,
-        skillId: target.groupId,
-      })
-      skillInventory.setData(nextInventory)
-      skillVersionReport.invalidate()
-      homeSummaryResource.invalidate()
-      setRuntimeSkillRemoveTarget(null)
-      toast.success(t("organizations.skillManageRemoveRuntimeSuccess", { name: target.displayName }))
-    } catch (error) {
-      toast.error(t("organizations.skillManageRemoveRuntimeFailed", { error: errorMessage(error) }))
-    } finally {
-      setBusyAction(null)
-    }
-  }, [busyAction, homeSummaryResource, runtimeSkillRemoveTarget, skillInventory, skillService, skillVersionReport, t])
-
-  const installRuntimeSkills = React.useCallback(
-    async (skills: readonly { packageName: string; skillName: string }[]) => {
-      const targets = skills.filter((skill) => skill.packageName.trim() && skill.skillName.trim())
-      if (targets.length === 0 || busyAction) {
-        return
-      }
-
-      setBusyAction("installSkillBatch")
-      let installedCount = 0
-      let failedCount = 0
-      let firstError: unknown
-      try {
-        for (const skill of targets) {
-          try {
-            const nextInventory = await skillService.invoke("installRegistrySkill", {
-              packageName: skill.packageName,
-              skillId: skill.skillName,
-            })
-            skillInventory.setData(nextInventory)
-            installedCount += 1
-          } catch (error) {
-            failedCount += 1
-            firstError ??= error
-          }
-        }
-        homeSummaryResource.invalidate()
-        if (installedCount > 0) {
-          toast.success(t("organizations.skillManageInstallMissingSuccess", { count: installedCount }))
-        }
-        if (failedCount > 0) {
-          toast.error(
-            t("organizations.skillManageInstallMissingFailed", {
-              count: failedCount,
-              error: errorMessage(firstError),
-            }),
-          )
-        }
-      } finally {
-        setBusyAction(null)
-      }
-    },
-    [busyAction, homeSummaryResource, skillInventory, skillService, t],
-  )
-
-  const linkOrganizationSkill = React.useCallback(
-    async (input: OrganizationSkillLinkInput, options: { installRuntime: boolean }) => {
-      if (!selectedOrganizationSkills?.canManage) {
-        return
-      }
-
-      await selectedOrganizationSkills.addSkill({
-        packageName: input.packageName,
-        skillName: input.skillName,
-        version: input.version,
-        versionPolicy: "pinned",
-      })
-      if (options.installRuntime) {
-        const nextInventory = await skillService.invoke("installRegistrySkill", {
-          packageName: input.packageName,
-          skillId: input.skillName,
-        })
-        skillInventory.setData(nextInventory)
-        homeSummaryResource.invalidate()
-      }
-    },
-    [homeSummaryResource, selectedOrganizationSkills, skillInventory, skillService],
-  )
-
-  const addOrganizationSkillFromRecommendation = React.useCallback(
-    async (recommendation: ProviderSkillRecommendation, options: { installRuntime: boolean }) => {
-      if (!selectedOrganizationSkills?.canManage || busyAction) {
-        return
-      }
-
-      setBusyAction(`addSkill:${recommendation.packageName}:${recommendation.skillId}`)
-      try {
-        await linkOrganizationSkill(
-          {
-            packageName: recommendation.packageName,
-            skillName: recommendation.skillId,
-            version: recommendation.package.version,
-          },
-          options,
-        )
-        toast.success(t("organizations.skillManageAddSuccess"))
-      } catch (error) {
-        toast.error(errorMessage(error))
-      } finally {
-        setBusyAction(null)
-      }
-    },
-    [busyAction, linkOrganizationSkill, selectedOrganizationSkills?.canManage, t],
-  )
-
-  const addOrganizationSkillFromPackage = React.useCallback(
-    async (pkg: PublicSkillPackage, options: { installRuntime: boolean; skillName?: string }) => {
-      if (!selectedOrganizationSkills?.canManage || busyAction) {
-        return
-      }
-
-      const input = publicPackageLinkInput(pkg, options.skillName)
-      if (!input) {
-        toast.error(t("skills.discoverInstallNoSkill"))
-        return
-      }
-
-      setBusyAction(`addSkill:${input.packageName}:${input.skillName}`)
-      try {
-        await linkOrganizationSkill(input, options)
-        toast.success(t("organizations.skillManageAddSuccess"))
-      } catch (error) {
-        toast.error(errorMessage(error))
-      } finally {
-        setBusyAction(null)
-      }
-    },
-    [busyAction, linkOrganizationSkill, selectedOrganizationSkills?.canManage, t],
-  )
-
-  const addOrganizationSkillBatch = React.useCallback(
-    async (recommendations: readonly ProviderSkillRecommendation[], options: { installRuntime: boolean }) => {
-      if (!selectedOrganizationSkills?.canManage || recommendations.length === 0 || busyAction) {
-        return
-      }
-
-      const plan = planOrganizationSkillBulkLinks(recommendations, selectedOrganizationSkills.skills)
-      if (plan.linkable.length === 0) {
-        return
-      }
-
-      setBusyAction("addSkillBatch")
-      let linkedCount = 0
-      let failedCount = 0
-      let firstError: unknown
-      try {
-        for (const recommendation of plan.linkable) {
-          try {
-            await linkOrganizationSkill(
-              {
-                packageName: recommendation.packageName,
-                skillName: recommendation.skillId,
-                version: recommendation.package.version,
-              },
-              options,
-            )
-            linkedCount += 1
-          } catch (error) {
-            failedCount += 1
-            firstError ??= error
-          }
-        }
-        if (linkedCount > 0) {
-          toast.success(
-            options.installRuntime
-              ? t("organizations.skillManageBulkAddInstallSuccess", { count: linkedCount })
-              : t("organizations.skillManageBulkAddSuccess", { count: linkedCount }),
-          )
-        }
-        if (failedCount > 0) {
-          toast.error(
-            t("organizations.skillManageBulkAddFailed", {
-              count: failedCount,
-              error: errorMessage(firstError),
-            }),
-          )
-        }
-      } finally {
-        setBusyAction(null)
-      }
-    },
-    [busyAction, linkOrganizationSkill, selectedOrganizationSkills, t],
   )
 
   return (
